[SonataImportBundle] Extract test double to fixture for PHPStan
PHPStan level 5 cannot see properties declared on an anonymous class
when accessed through the BaseColumnExtractor return type. Promote the
call-tracking double to a named fixture (CallTrackingColumnExtractor)
with explicit $callCount and $lastValue properties, instantiated
directly in the tests so PHPStan resolves the concrete type.
